Migrate zeroconf advertisement into its own module and test
