+++ .pylintrc
@@ -20,7 +20,8 @@ confidence=
 # --enable=similarities". If you want to run only the classes checker, but have
 # no Warning level messages displayed, use "--disable=all --enable=classes
 # --disable=W".
-disable=bare-except
+disable=bare-except,
+  logging-fstring-interpolation
 
 [BASIC]
 good-names=i,
