+++ scripts/test
@@ -27,6 +27,8 @@ pytest tests/test_ctrl.py -vvv -k doubly_missing_config
 
 pytest tests/test_rest.py -vvv -k 'not _live'
 pytest tests/test_auth.py -vvv
+pytest tests/test_zeroconf.py -vvv
+
 
 # Live tests require some amplipi streams to be setup, which means real audio
 # sinks are required. These tests will not run in Github Actions.
